requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==4.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
moviepy==1.0.3
numpy==2.2.6
pillow==11.2.1
proglog==0.1.12
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
